sql server is properly coonected now. Going to do the following nextL 1. Make the password WAY more secure thinking 16-32 random ascii chars 2. Finish off role based authorization implementation 3. Flesh out the API and possibly add front-end Razor/Blazor pages (Would love to try out blazor) 4. Publish the beta live at a subset Build The Lanes

+++ BuildTheLanesAPI/Models/Project.cs
@@ -5,7 +5,7 @@
 
     public class Project
     {
-        public long ProjectNumber { get; set; }
+        public int ProjectNumber { get; set; }
         public string StartDate { get; set; }
         public string Status { get; set; }
         public string City { get; set; }
